# src/main/java/com/review/reviewManagementService/models/User.java
package com.review.reviewManagementService.models;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class User {

    private Long Id;

    private String userName;

}